% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sites.R
\name{export_sites}
\alias{export_sites}
\title{Export Sites}
\usage{
export_sites(
  output_dir = NA,
  site_type_a = NA,
  type_b = NA,
  export_csv = TRUE,
  export_shp = TRUE,
  export_kml = TRUE
)
}
\arguments{
\item{output_dir}{String. Local file path the an output directory on your computer.}

\item{site_type_a}{Sample Sites Type A. sf data frame returned from grouped_random_sample()}

\item{type_b}{Sample Sites of Type B and line segments. Object returned from road_proximity_sample().}

\item{export_csv}{Boolean. Should a csv file be exported to the output directory.}

\item{export_shp}{Boolean. Should a shp (shape file) file be exported to the output directory.}

\item{export_kml}{Boolean. Should a kml file be exported to the output directory.}
}
\value{
Populates the export directory with output files
}
\description{
Export sampling sites for field programs
}
\details{
Define an output directory to export field sampling sites. Choose from available output formats. Note that when exporting to KML and SHP file format the data will be re-projected to latitude/longitude (EPSG: 4326).
}
\examples{
\dontrun{
}


}
