% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_alpine_bcfwa.R
\name{remove_alpine_bcfwa}
\alias{remove_alpine_bcfwa}
\title{Remove Alpine Reaches BCFWA}
\usage{
remove_alpine_bcfwa(strm = NA, elevation_threshold = 1000)
}
\arguments{
\item{strm}{A streamlines dataset of class \code{sf} from the BCFWA.}

\item{elevation_threshold}{Numeric. Elevation threshold for alpine areas.
Stream reaches above this threshold will be omitted from the analysis and
clipped to the threshold.}
}
\value{
An streamlines dataset of class \code{sf} with alpine areas removed
}
\description{
Remove alpine areas based on elevation parameters embedded within
BCFWA stream network. Choose a elevation threshold relevant to your
area of interest.
}
\details{
The BCFWA contains elevation data for each stream reach.
The z geometry (or elevation) is embedded within the stream line geometry
data and can be used to filter out or trim segments that are located
above a given elevation threshold. Note that converting the streamline
data from .gdb to .shp through intermediate data processing will result
in a loss of the z geometry. It is therefore recommended that users stick
to using .gdb or .gpkg file formats to preserve the elevation data.
}
\examples{
\dontrun{

library(wsep.t2)

# Trim stream reaches above 200m
data(TsolumStreams)
# Trim segments to an eelvation threshold
strm <- remove_alpine_bcfwa(strm = TsolumStreams,
 elevation_threshold = 200)
strm_plot <- sf::st_zm(strm)
plot(sf::st_geometry(TsolumStreams))
plot(sf::st_geometry(strm_plot), add = TRUE, col = "red")

# Trim stream reaches above 1200m
data(SpiusStreams)
# Trim segments to an eelvation threshold
strm <- remove_alpine_bcfwa(strm = SpiusStreams,
 elevation_threshold = 1200)
strm_plot <- sf::st_zm(strm)
plot(sf::st_geometry(SpiusStreams))
plot(sf::st_geometry(strm_plot), add = TRUE, col = "red")
}


}
