% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_frame_summary.R
\name{sample_frame_summary}
\alias{sample_frame_summary}
\title{Sample Frame Summary Table}
\usage{
sample_frame_summary(
  strm = NA,
  constrained_strm = NA,
  roads = NA,
  stream_order = "STREAM_ORDER",
  summary_type = "stream_lengths"
)
}
\arguments{
\item{strm}{A streamline data set of class \code{sf} from the BCFWA. Run \code{utm_projection()} first to ensure projection is in meters.}

\item{constrained_strm}{A the constrained streamline data set of class \code{sf}. Generally this layer is produced as a derivative of the \code{strm} layer after constraining NCD, excluding lentic areas and removing alpine areas.}

\item{roads}{A road network data set of class \code{sf}. Run \code{utm_projection()} first to ensure projection is in meters.}

\item{stream_order}{String. Column name for stream order. Defaults to \code{STREAM_ORDER}. Rename as needed. Stream order is used to define the sampling strata.}

\item{summary_type}{String. Which summary table should be produced? Options include \code{stream_lengths} for a summary of stream lengths by strata or \code{stream_crossings} for summary of stream crossings by strata. Rename as needed. Stream order is used to define the sampling strata.}
}
\value{
A summary table according to the \code{summary_type}.
}
\description{
Produce a summary table of the sampling frame.
}
\details{
It can be convenient to produce a summary table of the entire sampling frame by strata. This function summarizes the total length of streams by strata and the total count of stream crossings by strata.
}
