% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpiusRoads-data.R
\docType{data}
\name{SpiusRoads}
\alias{SpiusRoads}
\title{Spius Roads}
\format{
A road lines data set of class \code{sf}.
\describe{
\item{TRANSPORT_LINE_ID}{A unique identifier for each road line segment}
}
}
\usage{
data(SpiusRoads)
}
\description{
Sample road lines dataset for the Spius River watershed.
Used in a tutorial for the \code{wsep.t2}
(WSEP Tier II) R-package.
}
\examples{
\donttest{
require(sf)
data(SpiusRoads)
head(SpiusRoads)
# Plot the streams with stream order
plot(st_geometry(SpiusRoads))
}
}
\keyword{datasets}
