% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strm_grts.R
\name{strm_grts}
\alias{strm_grts}
\title{Stream GRTS Sample}
\usage{
strm_grts(n = 20, strm = NA, stream_order = "STREAM_ORDER")
}
\arguments{
\item{n}{Sample size n for each stratum. Defaults to 20.}

\item{strm}{A streamline data set of class \code{sf} from the BCFWA. Run \code{utm_projection()} first to ensure projection is in meters.}

\item{stream_order}{String. Column name for stream order. Defaults to \code{STREAM_ORDER}. Rename as needed.}
}
\value{
A object of class \code{sf} of sample features.
}
\description{
Generate a GRTS sample from stream segments.
}
\details{
Generate a GRTS sample from stream segments.
Part of Component 3 – Riparian. Site Type C.
GRTS Generalized Random Tesselation Stratified sampling is achieved using the \code{grts} function in the \code{spsurvey} package.
\itemize{
\item{Generate a GRTS sample from the stream segments for each of the two strata (<3rd order vs. ≥ 3rd order.}
\item{Create a field checklist with at least the key fields}
}
}
