% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsolumRoads-data.R
\docType{data}
\name{TsolumRoads}
\alias{TsolumRoads}
\title{Tsolum Roads}
\format{
A road lines data set of class \code{sf}.
\describe{
\item{TRANSPORT_LINE_ID}{A unique identifier for each road line segment}
}
}
\usage{
data(TsolumRoads)
}
\description{
Sample road lines dataset for the Tsolum River watershed near
Courtenay British Columbia. Used in the tutorial for the \code{wsep.t2}
(WSEP Tier II) R-package.
}
\examples{
\donttest{
require(sf)
data(TsolumRoads)
head(TsolumRoads)
# Plot the streams with stream order
plot(st_geometry(TsolumRoads))
}
}
\keyword{datasets}
