% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_projection.R
\name{utm_projection}
\alias{utm_projection}
\title{UTM Projection}
\usage{
utm_projection(data = NA, epsg = NA)
}
\arguments{
\item{data}{a spatial object class \code{sf}.}

\item{epsg}{optional EPSG (spatial reference number) for the local UTM zone.
If not set, the EPSG will be inferred from the extent of the data layer.}
}
\value{
A spatial object class \code{sf} projected to the appropriate UTM zone.
}
\description{
Project data layer to the local UTM zone
}
\details{
Since many of the \code{wsep.t2} processing operations require working with
real-world distances in units of meters, it is necessary to ensure that
the spatial data layers are projected to the a local UTM zone. This function
determines the appropriate UTM zone based on the extent of the data layer
and then transforms the data to that projection.
}
\examples{
\dontrun{
library(wsep.t2)
data(TsolumStreams)
strm <- utm_projection(data = TsolumStreams)
sf::st_crs(strm)$epsg
}


}
